[v14] buildbox: Update centos7 buildbox for EOL, using vault
CentOS 7 became EOL on 2024/06/30. `mirrorlist` and `mirror.centos.org`
became unavailable, so use `baseurl` and `vault.centos.org` instead,
updating all references in `/etc/yum.repos.d/*`.

The software collection stuff (`centos-release-scl-rh`) is on a
different path on `vault.centos.org` for `aarch64` so also update those
paths.

Apply these changes multiple times as installing `centos-release-scl-rh`
installs a yum repo file with the old paths, as does installing
`centos-release-scl`.

This backport is a little more involved than the original master PR as
`branch/v14` has old Dockerfiles here - there is no single base stage to
modify - a new `base` stage was added with the common stuff and the
other stages build from that instead of `centos:7`.

Also added is a `syntax` line to the two touched dockerfiles. This is
needed for here-documents to work (`<<EOF`) as the version of docker on
our GitHub runners is too old to understand that. The `syntax` line says
to use a different dockerfile parser to the built-in one.