[v14] buildbox: Remove package installation duplication
The previous commit added a new stage to the centos7 Dockerfiles, but
left package installation duplication in place. This commit removes the
packages installed in the base image from the subsequent images.

This has been done as a separate commit so that it can be reverted if
necessary. The first commit is the minimum needed, but leaves things a
little messy. This cleans up the mess but in case it breaks something,
this commit can be reverted.