[v16] rdp: Change rdpclient type casts from ulong to uintptr_t
Change the C type casts of `c.handle` from `C.ulong` to `C.uintptr_t`.
This is needed to compile on 32-bit architectures.

We do not currently build in the RDP client for 32-bit architectures but
it will be possible with a crosstool-NG based buildbox and is otherwise
of no detriment.

Co-authored-by: Jakub Nyckowski <jakub.nyckowski@goteleport.com>