Remove `aws-sdk-go/aws/awsutil` package (#43963)
We are migrating away from aws sdk v1.
This PR removes the usage of the awsutil package.

Changes include removing a helper method that was not ported to V2.
This method would copy from one struct into another one, using their
field names.
That's useful when we want to mutate a single field (eg update the ECS
image version).
However, that helper method was not ported.
Instead, we are now using json marshal/unmarshal to the same effect.