Fix redirect checker
The checker assumed that values of the `source` field in each redirect
configuration begin with `/docs/`, but they do not. Remove `/docs/` from
docs site URLs in the target directory before checking them against
configured redirects.

Fix an issue with the first item in the output of the program
always being `undefined` by correctly concatenating the result arrays.

Finally, fix a bug in which the checker would not accept redirects for
docs URLs with fragments.