ci: Add a 'run-name' to build-linux workflow (#4602)
Add a `run-name` field to the `build-linux` workflow so that a list of
parameterized runs do not all list as "Build Release on Linux" without
any idea of what the parameters are for the run.

The release-linux workflow was folded into the build-linux workflow
recently, and the former removed. However it had a `run-name` field
which had the parameters of the run but that was not moved across.

Source-commit: gravitational/teleport.e@ec132c173a6443244a9aa662e109c66ce7ddfeae
Source-pull-requests: gravitational/teleport.e#4602