build: Remove bogus cross-compiler setting (#4601)
Remove the settings for arm64 and arm setting the `CC` var for a cross
compiler. These settings never apply because the `$(IS_NATIVE_BUILD)`
variable is set to either `"yes"` or `"no"` and is never empty. The
`$(if ..)` expression removed would always evaluate to true.

The reason things still worked is that the `CC` var is set correctly in
the main Makefile and exported to the environment before calling the
enterprise Makefile, so it has the correct `CC` var by then.

Upcoming changes will invalidate this setting anyway, even if it was
set, so just remove it making it a bit cleaner to add better cross
compilation.

Source-commit: gravitational/teleport.e@7d737fb52ff3bf7c0ff1e82944f3967642e8137f
Source-pull-requests: gravitational/teleport.e#4601