build: Simplify the vars that target minimum macOS version (#4615)
Simplify how we tell rust/xcode/go which minimum version of macOS to
target. Xcode defines the environment variable
`MACOSX_DEPLOYMENT_TARGET` that contains the version of macOS to target.
llvm has adopted this too, so cargo/rust supports this, meaning we no
longer need to use `RUSTFLAGS` with a linker flag. This is much simpler
as we need to be careful not to set this flag when targetting wasm for
the UI (ironrdp). When building Go with CGO, it uses the system linker,
which is xcode, so we can rely on the `MACOSX_DEPLOYMENT_TARGET` var
there too, allowing us to get rid of the `CGO_LDFLAGS` var. We do still
need to set a linker flag on `CGO_CFLAGS` though, from testing.

Source-commit: gravitational/teleport.e@74e55e9ab7d3d12036574dc4d096a5e32160a491
Source-pull-requests: gravitational/teleport.e#4615