ci: Add permissions to release-mac workflow, lint fix (#4613)
Add permissions to the `release-mac` workflow when calling the
`build-mac` and `build-mac-universal` workflows. This recently changed
as the build workflows now upload to AWS S3 and need permissions to do
that. These perms normally come from `tag-build` but when doing just a
mac build, the `release-mac` workflow was failing with:

    The nested job 'build-release-artifacts' is requesting 'id-token: write',
    but is only allowed 'id-token: none'.

While we're here, fix a lint error now that we have actionlint as a CI
check.

Source-commit: gravitational/teleport.e@610b24c3eb90e5f91e88e35c608bc5356056dd3b
Source-pull-requests: gravitational/teleport.e#4613