[sec_scan][8] expose `backend.Item -> devicepb.Device` conversion func (#4597)
This PR exposes a function that converts a `backend.Item` into a `*devicepb.Device` object.
This is required because device trust service stores the object in the backend using an internal represenation that's not public.
Since the db events code needs to unmarshal backend objects when a new put/delete request is received in the database watcher, we need to expose the conversion logic.

This PR is part of https://github.com/gravitational/access-graph/issues/637.
This PR requires https://github.com/gravitational/teleport/pull/43905.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

Source-commit: gravitational/teleport.e@6e09dabb4dc033a74e7abcc4988a87152463484d
Source-pull-requests: gravitational/teleport.e#4597