Print accurate status in tsh login output (#44043)
If tsh login was executed while the user had a valid session the
resulting status shown to the user included data from the prior
session. The root cause was the profile being passed to
printLoginInformation was the old. Any profiles retrieved prior
to calling (TeleportClient)SaveProfile should be discarded and
the profile should be retrieved again to reflect the new session.

Closes https://github.com/gravitational/teleport/issues/33732.