Fixes a bug in the kube-agent-updater breaking private image resolution
Cosign was receiving the registry options contianing the auth keychain,
but its image resolution step was not. This caused the image resolution
to fail. This went undetected because we tested on already digested
images, so the resolution never happened.

This commit fixes the issue, introduces new tests covering both
authenticated registries and image resoluition.

This commit also improves the error logging. When logging trust errors,
we nmow log the original error to provide more context. This also adds
more debug logging.