Update rebuild-teleport-oci-distroless.yml (#4605)
* Update rebuild-teleport-oci-distroless.yml

Use build-prod environment for scheduled oci image rebuilds.

* ci: Use parse-version in rebuild-teleport-oci-distroless workflow

Change the hacky bash to our `parse-version` action in the
`rebuild-teleport-oci-distroless` workflow which removes a bunch of
actionlint errors present in the bash code. `parse-version` was
developed after this bash code was written, and was always meant to be
replaced. The time has come.

---------

Co-authored-by: Cam Hutchison <camh@goteleport.com>

Source-commit: gravitational/teleport.e@91a7731634520cf7619944146188fc47ce4ad0eb
Source-pull-requests: gravitational/teleport.e#4605