[v16] Add implementation of VNet daemon (#44265)
* Update the daemon section of RFD

This reflects how the daemon actually end up being implemented.

* Add info about FIDO2 to tshdev/README.md

* Call the daemon from the client

The next commit is going to actually add the implementation of
the daemon service.

* Rename AdminSubcommand to AdminSetup

This function is going to be executed by both tsh vnet-admin-setup and
tsh vnet-daemon.

"Admin subcommand" in the context of lib/vnet/setup.go could mean either
of those two subcommands, but what we're really interested in is the
process itself.

* Extract execAdminProcess args to struct

* Pass home path explicitly to OS configurator and daemon.Config

We're not able to dynamically set a custom TELEPORT_HOME env var when
launching the daemon, so we have to pass it explicitly through XPC.

* Implement daemon subcommand

* Synchronize reading data from daemonService

* Validate config before use, not after creation

* daemon.Config.CheckAndSetDefaults: Use switch

* Improve how Go errors are created

* Adjust & add godocs, add missing stdlib.h include

* Fix includes and imports

* Make DaemonStart a noop if daemon was already started

* Do not validate config on the client side

* Remove asserts