build: Also look in C_INCLUDE_PATH for pam_appl.h
When looking for the PAM header `pam_appl.h` to determine if PAM should
be enabled in teleport, first look in the directories specified in the
`C_INCLUDE_PATH` environment variable before checking the system header
directories. The buildbox-ng has per-architecture include directories
and configures `C_INCLUDE_PATH` to find the correct headers. Fall back
to the existing default directories if not found in `C_INCLUDE_PATH`.