Remove redirects from non-default branches
Since the docs sitemap only applies to the default docs site version,
Google only indexes that version. External links in the Teleport product
do not refer to specific versions of the docs site. As a result, it
makes sense to only maintain redirects on the default version of the
docs site.

Beyond only including the redirects we need, this approach has the
advantage of reducing our overall Vercel route count so we can add more
redirects for the default branch. This is necessary for the next phase
of the docs reorganization.