Update lib/events/dynamoevents to use aws-sdk-go-v2
This is a continuation of converting dynamodb components to use the
latest version of the sdk that was started in
https://github.com/gravitational/teleport/pull/44356.

This should have feature parity with the existing backend except
for prometheus metrics. In an attempt to isolate the changes here
the metrics are omitted for the time being and will be added in a
follow up.

In addition, a few of the events test suite cases were updated to
be more reliable when testing against a real backend.