Add observability for github.com/aws/aws-sdk-go-v2
Implements middleware that capture prometheus metrics for interactions
with the AWS API. The dynamo and s3 metrics wrappers over the
aws-sdk-go-v1 API have been converted as well as a new set of
metrics that are AWS service agnostic. This should allow future
adoption of AWS metrics to be painless and not require writing a
middleware per service. Additionally, observability was extended by
leverating otelaws to add tracing to dynamodb requests.