[v16] build: Find feature headers outside system include dir
* build: Use pkg-config to find bpf headers and libraries

Use `pkg-config` / `pkgconf` to find the specific version of `libbpf`
that we want to link to as well as the dependent libraries such as
`libelf`, `libz`, and sometimes `libzstd`. This resolves the variable
dependency on `libzstd` depending on the base OS.

We still take preference of `/lib/libbpf-$(LIBBPF_VER)` if it exists as
it does for the standard buildbox, but that is deprecated and will be
removed when we move to the new buildbox.

If there is no `libelf.pc` config, fall back to hard-coded libraries as
it was previously. Again, this is for use with our existing CentOS 7
buildbox.

This cleans up the BPF detection in `common.mk` as it had some
head-scratching failure modes when clang or llvm-strip were not found.
The logic for enabling BPF is clearer now and safer.

* build: Also look in C_INCLUDE_PATH for pam_appl.h

When looking for the PAM header `pam_appl.h` to determine if PAM should
be enabled in teleport, first look in the directories specified in the
`C_INCLUDE_PATH` environment variable before checking the system header
directories. The buildbox-ng has per-architecture include directories
and configures `C_INCLUDE_PATH` to find the correct headers. Fall back
to the existing default directories if not found in `C_INCLUDE_PATH`.