Update lint action to catch any untidy go module
Due to integrations/terraform and integrations/event-handler
importing teleport and replacing with ../.. any updates to
go.mod need to be kept in sync with the integrations modules.
The go mod tidiness check in the lint step only compared teleport
and the api modules, which allowed things to get out of sync and
possibly break integrations builds. The tidiness check now validates
that any go module in the repository is tidy.