Update lib/backend/dynamo to use aws-sdk-go-v2 (#44356)
In addition to converting the backend, this also removes the
dependency of lib/backend/dynamo from lib/dynamoevents. The events
backend used some helpers defined in the lib/backend/dynamo to
configure the dynamodb table. In the interest in converting one
dynamodb package at a time lib/backend/dynamo/configure was removed
and the shared code copied to both packages.

This should have feature parity with the existing backend except
for prometheus metrics. In an attempt to isolate the changes here
the metrics are omitted for the time being and will be added to
both the state and events backends after they have been converted.