[v16] Bump pgx/v5 to v5.6.0
This version includes a change which updates the default dialer's
keepalive configuration. This reduces the time for a dead connection
to be cleaned up from 50m to 2m30s.

https://github.com/jackc/pgx/commit/24c0a5e8ff740778bbed9810629cb83c9ea18bf9

The impact of a dead connection is that a query will hang until the connection
is cleaned up. This can cause agents to hang on heartbeats and disappear from
the backend.