[v16] Adds detailed Okta status to Plugin Resource
Backports #44431

Once we collect the Okta usage statistics we need a durable place to store
them. This patch adds a `Details` block to the Okta plugin status in order
provide such a location. As an extra bonus, we already know that the plugin
status can be reached from both the running plugin and the UI code making it
easier plumb the data through in a subsequent PR.